Modified configure.ac to retain CPPFLAGS at final GEOS subsitution
Line 527 of configure.ac was replacing all the accumulated CPPFLAGS
with only those for the GEOS package.  Adding the ${CPPFLAGS} to the
front of the final substitution so that they are included along with
the GEOS PKG_CPPFLAGS seems to work.

All of the changes to configure are a result of running autoreconf.